requirements.txt
Django==5.2.4
django-recaptcha==4.1.0
asgiref==3.9.0
sqlparse==0.5.3
tzdata==2025.2
python-dotenv==1.0.0
django-csp==3.7
django-axes==6.1.1
django-ratelimit==4.1.0
whitenoise==6.6.0
gunicorn==21.2.0
psycopg2-binary==2.9.9
dj-database-url==2.1.0
polib==1.2.0